Extract asset-related methods from DagsterInstance into AssetMixin
This refactoring moves approximately 80 asset-related methods from the DagsterInstance
class into a new AssetMixin class to improve code organization and maintainability.
The change follows the existing pattern used by other mixins like SettingsMixin,
RunsMixin, and DomainsMixin.

Key changes:
- Created new AssetMixin class containing all asset-related methods
- Updated DagsterInstance to inherit from AssetMixin for backward compatibility
- Cleaned up imports in instance.py
- Maintained all existing public APIs
